refactor(http): extract environment editor state into a composable
